[refactor🔨] : JPA Entity @Column 제거, Pk 컨벤션에 맞추어 수정, 테스트 관리

+++ src/main/java/shop/gaship/gashipshoppingmall/member/repository/MemberRepository.java
@@ -3,5 +3,5 @@ package shop.gaship.gashipshoppingmall.member.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import shop.gaship.gashipshoppingmall.member.entity.Member;
 
-public interface MemberRepository extends JpaRepository<Member, Long>, MemberRepositoryCustom {
+public interface MemberRepository extends JpaRepository<Member, Integer>, MemberRepositoryCustom {
 }
